% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr2-package.R
\docType{package}
\name{bkmr2-package}
\alias{bkmr2-package}
\title{The 'bkmr2' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1. https://mc-stan.org
}
