% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr2.R
\name{bkmr2}
\alias{bkmr2}
\title{Bayesian Kernel Machine Regression}
\usage{
bkmr2(
  formula,
  data,
  P,
  sigma_a = 1,
  sigma_b = 3,
  phi_a = 1,
  phi_b = 3,
  est_phi = FALSE,
  ...
)
}
\arguments{
\item{formula}{formula to specify design matrix X}

\item{data}{data for design matrix X}

\item{P}{covariance distance matrix}

\item{sigma_a}{sigma normal prior mean}

\item{sigma_b}{sigma normal prior scale}

\item{phi_a}{phi normal prior mean}

\item{phi_b}{phi normal prior scale}

\item{est_phi}{boolean value indicating whether phi should be estimated}

\item{...}{optional arguments passed to \code{\link[rstan]{sampling}}}
}
\value{
bkmreg object
}
\description{
Bayesian Kernel Machine Regression
}
