% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmreg-methods.R
\name{bkmreg-methods}
\alias{bkmreg-methods}
\alias{nsamples.bkmreg}
\title{Methods for bkmreg objects}
\usage{
\method{nsamples}{bkmreg}(object, ...)
}
\arguments{
\item{object}{bkmreg object}

\item{...}{optional arguments}
}
\description{
Explanatory text
}
