% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{h_dist}
\alias{h_dist}
\alias{h_dist.bkmreg}
\title{|h| as a function of distance in P}
\usage{
h_dist(x, num_sample = 3)

\method{h_dist}{bkmreg}(x, num_sample = 3)
}
\arguments{
\item{x}{bkmreg object}

\item{num_sample}{number of sample indices to show}
}
\description{
|h| as a function of distance in P
}
\section{Methods (by class)}{
\itemize{
\item \code{bkmreg}: bkmreg h dist plot
}}

