% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bkmreg}
\alias{plot.bkmreg}
\title{Plotting function for bkmreg objects}
\usage{
\method{plot}{bkmreg}(x, plotfun = "h_dist", ...)
}
\description{
@aliases plot
@param x bkmreg object
@param plotfun one of c("h_dist","ppc")
@param ... optional arguments for \code{plotfun}
@export
}
