% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ppc}
\alias{ppc}
\alias{ppc.bkmreg}
\title{Posterior Predictive Checks}
\usage{
ppc(x, num_reps = 10)

\method{ppc}{bkmreg}(x, num_reps = 10)
}
\arguments{
\item{x}{bkmrreg object}

\item{num_reps}{number of replications to show}
}
\description{
Posterior Predictive Checks
}
\section{Methods (by class)}{
\itemize{
\item \code{bkmreg}: bkmr posterior predictive check plot
}}

