% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bkmreg}
\alias{plot.bkmreg}
\title{Plotting function for bkmreg objects}
\usage{
\method{plot}{bkmreg}(x, plotfun = "h_dist", ...)
}
\arguments{
\item{x}{bkmreg object}

\item{plotfun}{one of c("h_dist","ppc")}

\item{...}{optional arguments for \code{plotfun}}
}
\description{
Plotting function for bkmreg objects
}
